bharatguard/requirements.txt
langgraph==0.2.20
langchain==0.3.0
langchain-community
ollama
chromadb
gitpython
pydantic==2.10
python-dotenv
reportlab
streamlit
openai-whisper
fastapi
uvicorn
